add and remove item implemented
Functions to add and remove items updated.
Now, it can add items passing parameters needed to user be able to delete the row he want, not only the last row. Each row has a remove button that refers to itself

It means, for example, that if the user add 5 items and need to remove the third one, he will not need to remove and add items 4 and 5 again.

- minors
- - externalized js functions
- - routes turned in relative ones