Some pivot on sale register functionality
Tables, functions, endpoints and objects were created to improve the way the products are registered.

Instead of only keep it on frontend and submit everithing in the end, each step will be recorded on database, through the functions holdItem and openReceipt.

openReceipt will create a new receipt when the user set a costumer to sell, and set status 1 to it.
If this costumer change, the receipt_id ill be keeped.

After this, each item will be registered on a table that have receipt+products relationship, and if the user remove a field group that is a item, this item will be deleted from the base.

After all, when submit, the receipt status will be setted to 2, and be no longer avaliable to input, update or delete a new item.